+++ tests/e2e/conftest.py
@@ -1,6 +1,7 @@
 import logging
 import os
 import uuid
+from concurrent.futures import Executor
 from datetime import (
     datetime,
     timezone,
@@ -9,10 +10,18 @@ from datetime import (
 from neptune_scale import Run
 from pytest import fixture
 
+from neptune_api import AuthenticatedClient
+from neptune_api.credentials import Credentials
 from neptune_fetcher import (
     ReadOnlyProject,
     ReadOnlyRun,
 )
+from neptune_fetcher.api.api_client import (
+    create_auth_api_client,
+    get_config_and_token_urls,
+)
+from neptune_fetcher.internal.composition import concurrency
+
 
 # NOTE
 # The fixtures below assume that we're testing on a "many metrics" project
@@ -20,6 +29,21 @@ from neptune_fetcher import (
 # contains ONLY data generated by the script.
 # If this is not the case, the tests will fail, as there are hard assumptions
 # around names and values of the metadata.
+
+
+API_TOKEN_ENV_NAME: str = "NEPTUNE_API_TOKEN"
+
+
+@fixture(scope="session")
+def client() -> AuthenticatedClient:
+    api_token = os.getenv(API_TOKEN_ENV_NAME)
+    credentials = Credentials.from_api_key(api_key=api_token)
+    config, token_urls = get_config_and_token_urls(credentials=credentials, proxies=None)
+    client = create_auth_api_client(
+        credentials=credentials, config=config, token_refreshing_urls=token_urls, proxies=None
+    )
+
+    return client
 
 
 @fixture(scope="session", autouse=True)
@@ -37,6 +61,11 @@ def cleanup_logging_handlers():
     finally:
         logger = logging.getLogger("neptune")
         logger.handlers.clear()
+
+
+@fixture(scope="module")
+def executor() -> Executor:
+    return concurrency.create_thread_pool_executor()
 
 
 @fixture(scope="module")
