+++ tests/e2e/internal/retrieval/test_search.py
@@ -20,7 +20,7 @@ from neptune_fetcher.internal.retrieval.search import (
     ExperimentSysAttrs,
     fetch_experiment_sys_attrs,
 )
-from tests.e2e.alpha.internal.data import (
+from tests.e2e.internal.data import (
     FLOAT_SERIES_PATHS,
     PATH,
     STRING_SERIES_PATHS,
