+++ tests/unit/internal/retrieval/test_util.py
