+++ tests/unit/internal/test_output_format.py
